Resolve dependencies needed to build wheels
* Clean up MANIFEST.in

* Add PyYAML as a setup requirement

The PyYAML package is needed to create a source distribution in a clean
environment. For this setup requirement to be installed for the command,
a custom sdist command is needed so it is executed _after_ setup
requirements have been installed.

* Reduce dependencies needed for pickle file

Since the library is needed to build the pickle file when installing
itself, it would be good to only require the parts needed to build the
pickle file.

Signed-off-by: Craig Northway <cnorthway@codeaurora.org>